Automatically backed up by Learn

+++ lib/create.sql
@@ -3,6 +3,6 @@ CREATE TABLE projects (
    PRIMARY KEY,
   title TEXT,
   category TEXT,
-  funding_goal NUMERIC
+  funding_goal INTEGER
 
 );
